Create manifest.json
To use as a faux plugin for dev channel edition of Google Chrome. Create a folder somewhere, place the manifest and the script files in it. Open the the extensions management page chrome://extensions/ Click on developer mode checkbox. Click on the load unpacked extension and navigate to the folder where you placed the files.